requirements.txt
requests
gpsd-py3
pymodbus
